Revert "Add Postgresql User Store"
This reverts commit 746fe0be02c9edbd5b4383c02adbc22c5d89e1bf.
