+++ function/Heartbeat.cs
@@ -27,7 +27,7 @@ namespace Surfrider.PlasticOrigins.Backend.Mobile
         {
             log.LogInformation("Heartbeat request");
 
-            var userStorageAvailability = await _userStore.CheckAvailability();
+            var userStorageAvailability = _userStore.CheckAvailability();
 
             return (ActionResult)new OkObjectResult(
                 new
